[PM-21359] Migrate StringExtensions to `ui` module
Move StringExtensions to the shared `ui` module.
