+++ app/src/main/java/com/x8bit/bitwarden/ui/vault/feature/importlogins/ImportLoginsScreen.kt
@@ -39,6 +39,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.bitwarden.core.annotation.OmitFromCoverage
 import com.bitwarden.ui.platform.base.util.EventsEffect
+import com.bitwarden.ui.platform.base.util.toAnnotatedString
 import com.x8bit.bitwarden.R
 import com.x8bit.bitwarden.ui.platform.base.util.bitwardenBoldSpanStyle
 import com.x8bit.bitwarden.ui.platform.base.util.standardHorizontalMargin
