+++ authenticator/src/main/kotlin/com/bitwarden/authenticator/ui/platform/components/stepper/BitwardenStepper.kt
@@ -6,12 +6,12 @@ import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.text.input.KeyboardType
 import com.bitwarden.authenticator.R
-import com.bitwarden.authenticator.ui.platform.base.util.ZERO_WIDTH_CHARACTER
-import com.bitwarden.authenticator.ui.platform.base.util.orNullIfBlank
 import com.bitwarden.authenticator.ui.platform.components.field.BitwardenTextFieldWithActions
 import com.bitwarden.authenticator.ui.platform.components.icon.BitwardenIconButtonWithResource
 import com.bitwarden.authenticator.ui.platform.components.model.IconResource
 import com.bitwarden.authenticator.ui.platform.components.util.rememberVectorPainter
+import com.bitwarden.ui.platform.base.util.ZERO_WIDTH_CHARACTER
+import com.bitwarden.ui.platform.base.util.orNullIfBlank
 
 /**
  * Displays a stepper that allows the user to increment and decrement an int value.
