+++ ui/src/test/kotlin/com/bitwarden/ui/platform/base/util/StringExtensionsTest.kt
@@ -1,10 +1,10 @@
-package com.x8bit.bitwarden.ui.platform.base.util
+package com.bitwarden.ui.platform.base.util
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
-import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertNull
 
 class StringExtensionsTest {
 
@@ -206,5 +206,50 @@ class StringExtensionsTest {
         val expected = "https://$uri"
         val actual = uri.prefixHttpsIfNecessaryOrNull()
         assertEquals(expected, actual)
+    }
+
+    @Test
+    fun `orNullIfBlank returns null for a null String`() {
+        assertNull((null as String?).orNullIfBlank())
+    }
+
+    @Test
+    fun `orNullIfBlank returns null for an empty String`() {
+        assertNull("".orNullIfBlank())
+    }
+
+    @Test
+    fun `orNullIfBlank returns null for a blank String`() {
+        assertNull("      ".orNullIfBlank())
+    }
+
+    @Test
+    fun `orNullIfBlank returns the original value for a non-blank String`() {
+        assertEquals("test", "test".orNullIfBlank())
+    }
+
+    @Test
+    fun `orNullIfBlank returns the original value for a zero-width space String`() {
+        assertEquals("\u200B", "\u200B".orNullIfBlank())
+    }
+
+    @Test
+    fun `orZeroWidthSpace returns null for a null String`() {
+        assertEquals("\u200B", null.orZeroWidthSpace())
+    }
+
+    @Test
+    fun `orZeroWidthSpace returns zero-width space for an empty String`() {
+        assertEquals("\u200B", "".orZeroWidthSpace())
+    }
+
+    @Test
+    fun `orZeroWidthSpace returns zero-width space for a blank String`() {
+        assertEquals("\u200B", "      ".orZeroWidthSpace())
+    }
+
+    @Test
+    fun `orZeroWidthSpace returns the original value for a non-blank string`() {
+        assertEquals("test", "test".orZeroWidthSpace())
     }
 }
